+++ app/src/main/java/com/example/clicker/presentation/home/HomeView.kt
@@ -162,7 +162,6 @@ fun ValidationView(
         hapticFeedBackError={hapticFeedBackError()},
         lowPowerModeActive=lowPowerModeActive,
         changeLowPowerMode={newValue ->streamViewModel.changeLowPowerModeActive(newValue)},
-        offlineStreams = homeViewModel.offlineFollowedStreams.value,
 
     )
