--- app/src/main/java/com/example/clicker/presentation/sharedViews/SharedComponents.kt
@@ -1,492 +1,0 @@
-package com.example.clicker.presentation.sharedViews
-
-import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.Icon
-import androidx.compose.material.ScaffoldState
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Check
-import androidx.compose.material3.CardDefaults
-import androidx.compose.material3.DrawerState
-import androidx.compose.material3.DrawerValue
-import androidx.compose.material3.ElevatedCard
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.ModalDrawerSheet
-import androidx.compose.material3.ModalNavigationDrawer
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Switch
-import androidx.compose.material3.SwitchDefaults
-import androidx.compose.material3.Text
-import androidx.compose.material3.rememberDrawerState
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.painter.Painter
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import com.example.clicker.R
-import com.example.clicker.util.NetworkAuthResponse
-import kotlinx.coroutines.launch
-
-/**
- * SharedComponents represents the most used and most stable versions of components used throughout this application
- * **/
-object SharedComponents {
-
-    /**
-     * NoDrawerScaffold is a [Scaffold] composable meant to be used to build a scaffold containing a [topBar],
-     * [bottomBar] and [content] but no drawer
-     * - a UI demonstration can be found [HERE](https://github.com/thePlebDev/Clicker/wiki/Shared-Components#NoDrawerScaffold)
-     *
-     * @param topBar a [ScaffoldTopBarScope] composable that will be displayed in the Scaffold's topBar
-     * @param bottomBar a [ScaffoldTopBarScope] composable that will be displayed in the Scaffold's bottomBar
-     * @param content a normal composable that will be displayed in the Scaffold's content section
-     *
-     * */
-    @Composable
-    fun NoDrawerScaffold(
-        topBar:@Composable ScaffoldTopBarScope.() -> Unit,
-        bottomBar:@Composable ScaffoldBottomBarScope.() -> Unit,
-        content:@Composable (contentPadding: PaddingValues,) -> Unit,
-
-        ) {
-        val topBarScaffoldScope = remember(){ScaffoldTopBarScope(35.dp)}
-        val bottomBarScaffoldScope = remember(){ScaffoldBottomBarScope(25.dp)}
-
-        Scaffold(
-            containerColor = MaterialTheme.colorScheme.primary,
-            topBar = {
-                Row(modifier = Modifier
-                    .fillMaxWidth()
-                    .background(MaterialTheme.colorScheme.primary)){
-                    with(topBarScaffoldScope){
-                        topBar()
-                    }
-                }
-
-            },
-            bottomBar = {
-                with(bottomBarScaffoldScope){
-                    bottomBar()
-                }
-            },
-
-            ) { contentPadding ->
-            content(contentPadding)
-
-        }
-    }
-
-    /**
-     * NoDrawerScaffold is a [Scaffold] composable meant to be used to build a scaffold containing a [topBar],
-     * [bottomBar] and [content] with a drawer
-     * - a UI demonstration can be found [HERE](https://github.com/thePlebDev/Clicker/wiki/Shared-Components#DrawerScaffold)
-     *
-     * @param scaffoldState a [ScaffoldState] which is used to open and close the Scaffold's drawer
-     * @param topBar a [ScaffoldTopBarScope] composable that will be displayed in the Scaffold's topBar
-     * @param bottomBar a [ScaffoldTopBarScope] composable that will be displayed in the Scaffold's bottomBar
-     * @param drawerContent a normal composable that will be displayed in the Scaffold's drawerContent
-     * @param content a normal composable that will be displayed in the Scaffold's content section
-     *
-     * */
-    @Composable
-    fun DrawerScaffold(
-        scaffoldState: ScaffoldState,
-        topBar:@Composable ScaffoldTopBarScope.() -> Unit,
-        bottomBar:@Composable ScaffoldBottomBarScope.() -> Unit,
-        drawerContent:@Composable () -> Unit,
-        content:@Composable (contentPadding: PaddingValues,) -> Unit,
-    ){
-        val topBarScaffoldScope = remember(){ScaffoldTopBarScope(35.dp)}
-        val bottomBarScaffoldScope = remember(){ScaffoldBottomBarScope(25.dp)}
-
-
-        androidx.compose.material.Scaffold(
-            scaffoldState = scaffoldState,
-            backgroundColor = MaterialTheme.colorScheme.primary,
-                topBar = {
-                    Row(modifier = Modifier
-                        .fillMaxWidth()
-                        .background(MaterialTheme.colorScheme.primary)){
-                        with(topBarScaffoldScope){
-                            topBar()
-                        }
-                    }
-                },
-            drawerContent = {
-                drawerContent()
-            },
-                bottomBar = {
-                    with(bottomBarScaffoldScope){
-                        bottomBar()
-                    }
-                },
-
-                ) { contentPadding ->
-                content(contentPadding)
-            }
-
-
-
-    }
-}
-
-
-@Composable
-fun ModViewScaffoldWithDrawer(
-    topBar:@Composable (showDrawer:()->Unit) -> Unit,
-    bottomBar:@Composable ScaffoldBottomBarScope.() -> Unit,
-    drawerState: DrawerState,
-    checkIndexAvailability:(Int)->Unit,
-    showError:Boolean,
-    autoModQueueChecked:Boolean,
-    changeAutoModQueueChecked:(Boolean)->Unit,
-
-    modActionsChecked:Boolean,
-    changeModActionsChecked:(Boolean)->Unit,
-    content:@Composable (contentPadding: PaddingValues,) -> Unit,
-
-    ) {
-    val topBarScaffoldScope = remember(){ScaffoldTopBarScope(35.dp)}
-    val bottomBarScaffoldScope = remember(){ScaffoldBottomBarScope(25.dp)}
-    val scope = rememberCoroutineScope()
-
-    ModalNavigationDrawer(
-        drawerState = drawerState,
-        drawerContent = {
-            ModalDrawerSheet {
-                ModViewDrawerContent(
-                    checkIndexAvailability={index -> checkIndexAvailability(index)},
-                    showError =showError,
-                    autoModQueueChecked = autoModQueueChecked,
-                    changeAutoModQueueChecked ={value ->changeAutoModQueueChecked(value)},
-                    modActionsChecked=modActionsChecked,
-                    changeModActionsChecked ={value ->changeModActionsChecked(value)}
-
-                )
-            }
-        },
-    ) {
-        Scaffold(
-            containerColor = MaterialTheme.colorScheme.primary,
-            topBar = {
-                Row(modifier = Modifier
-                    .fillMaxWidth()
-                    .background(MaterialTheme.colorScheme.primary)){
-
-                        topBar(
-                            showDrawer={
-                                scope.launch {
-                                    drawerState.open()
-                                }
-                            }
-                        )
-
-                }
-
-            },
-            bottomBar = {
-                with(bottomBarScaffoldScope){
-                    bottomBar()
-                }
-            },
-
-            ) { contentPadding ->
-            content(contentPadding)
-
-        }
-    }
-
-
-}
-
-@Composable
-fun ModViewDrawerContent(
-    checkIndexAvailability:(Int)->Unit,
-    showError:Boolean,
-    autoModQueueChecked:Boolean,
-    changeAutoModQueueChecked:(Boolean)->Unit,
-
-    modActionsChecked:Boolean,
-    changeModActionsChecked:(Boolean)->Unit,
-){
-    Box(modifier = Modifier.fillMaxSize()){
-        LazyColumn(
-            modifier = Modifier.padding(horizontal = 10.dp)
-        ) {
-            item {
-                ElevatedCardSwitchTextRow(
-                    "Chat",
-                    checkIndexAvailability={checkIndexAvailability(1)},
-                    painter = painterResource(id =R.drawable.keyboard_24),
-                )
-
-            }
-            item{
-                ElevatedCardSwitchRow(
-                    "AutoMod Queue",
-                    checkIndexAvailability={checkIndexAvailability(2)},
-                    painter = painterResource(id =R.drawable.mod_view_24),
-                    checked = autoModQueueChecked,
-                    changeChecked = {value -> changeAutoModQueueChecked(value)}
-                )
-            }
-            item{
-                ElevatedCardSwitchRow(
-                    "Mod actions",
-                    checkIndexAvailability={checkIndexAvailability(3)},
-                    painter = painterResource(id =R.drawable.clear_chat_alt_24),
-                    checked = modActionsChecked,
-                    changeChecked = {value -> changeModActionsChecked(value)}
-                )
-
-
-            }
-
-//            item{
-//                ElevatedCardExample(
-//                    Color.Yellow,
-//                    "Un-ban requests",
-//                    checkIndexAvailability={checkIndexAvailability(4)}
-//                )
-//            }
-//
-//            item{
-//                ElevatedCardExample(
-//                    Color.LightGray,
-//                    "Discord",
-//                    checkIndexAvailability={checkIndexAvailability(5)}
-//                )
-//            }
-//
-//            item{
-//                ElevatedCardExample(
-//                    Color.Cyan,
-//                    "Moderators",
-//                    checkIndexAvailability={checkIndexAvailability(6)}
-//                )
-//            }
-
-        }
-
-        if(showError){
-            Column(modifier = Modifier.align(Alignment.BottomCenter)) {
-                ErrorMessage(
-                    modifier = Modifier,
-                    message="Error! No space to place "
-                )
-                Spacer(modifier =Modifier.height(10.dp))
-            }
-        }
-
-    }
-
-}
-
-@Composable
-fun ElevatedCardSwitchRow(
-    text:String,
-    checkIndexAvailability: () -> Unit,
-    painter: Painter,
-    checked:Boolean,
-    changeChecked:(Boolean) ->Unit,
-){
-    Row(
-        modifier = Modifier.fillMaxWidth(),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.SpaceBetween
-    ){
-        ElevatedCardWithIcon(
-            text,
-            checkIndexAvailability={checkIndexAvailability()},
-            painter = painter
-        )
-
-        SwitchWithIcon(
-            checked = checked,
-            changeChecked ={value -> changeChecked(value)}
-        )
-    }
-}
-
-@Composable
-fun ElevatedCardSwitchTextRow(
-    text:String,
-    checkIndexAvailability: () -> Unit,
-    painter: Painter
-){
-    Row(
-        modifier = Modifier.fillMaxWidth(),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.SpaceBetween
-    ){
-        ElevatedCardWithIcon(
-            text,
-            checkIndexAvailability={checkIndexAvailability()},
-            painter = painter,
-
-            )
-
-        TextColumn(text="Notifications")
-    }
-}
-
-@Composable
-fun SwitchWithIcon(
-    checked:Boolean,
-    changeChecked:(Boolean) ->Unit,
-
-) {
-
-
-    Column(
-        modifier = Modifier.fillMaxWidth().height(90.dp).padding(top=13.dp),
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.Center
-    ) {
-
-        Switch(
-            checked = checked,
-            onCheckedChange = {
-                changeChecked(it)
-            },
-            thumbContent = if (checked) {
-                {
-                    Icon(
-                        imageVector = Icons.Filled.Check,
-                        contentDescription = null,
-                        modifier = Modifier.size(SwitchDefaults.IconSize),
-                        tint = Color.White
-                    )
-                }
-            } else {
-                null
-            },
-            colors = SwitchDefaults.colors(
-                checkedThumbColor = MaterialTheme.colorScheme.secondary,
-                uncheckedThumbColor = MaterialTheme.colorScheme.secondary,
-                checkedTrackColor = Color.DarkGray,
-                uncheckedTrackColor = Color.DarkGray,
-            )
-        )
-    }
-
-}
-
-@Composable
-fun TextColumn(
-    text:String
-) {
-
-
-    Column(
-        modifier = Modifier.fillMaxWidth().height(90.dp).padding(top=13.dp),
-        horizontalAlignment = Alignment.CenterHorizontally,
-    ) {
-        Text(
-            text = text,
-            color = Color.White,
-            fontSize = 20.sp
-        )
-    }
-
-}
-
-@Composable
-fun ElevatedCardWithIcon(
-    type:String,
-    checkIndexAvailability:()->Unit,
-    painter: Painter
-) {
-    Column() {
-        Spacer(modifier =Modifier.height(15.dp))
-        ElevatedCard(
-            colors = CardDefaults.cardColors(containerColor = Color.DarkGray),
-            elevation = CardDefaults.cardElevation(
-                defaultElevation = 6.dp
-            ),
-            modifier = Modifier
-                .size(width = 200.dp, height = 80.dp)
-                .clickable { checkIndexAvailability() }
-        ) {
-            Column(
-                modifier = Modifier.fillMaxSize(),
-                horizontalAlignment = Alignment.CenterHorizontally,
-                verticalArrangement = Arrangement.Center
-            ){
-                Text(
-                    text = type,
-                    color = Color.White,
-                    modifier = Modifier,
-                    fontSize = 20.sp
-                )
-                Icon(
-                    painter = painter,
-                    contentDescription ="",
-                    tint = Color.White,
-                    modifier = Modifier.size(30.dp)
-                )
-
-            }
-
-
-        }
-        Spacer(modifier =Modifier.height(15.dp))
-    }
-
-}
-@Composable
-fun ErrorMessage(
-    modifier: Modifier,
-    message:String,
-){
-
-    Row(
-        modifier = modifier
-            .clip(
-                RoundedCornerShape(20.dp)
-            )
-            .background(Color.Red)
-            .padding(vertical = 5.dp, horizontal = 10.dp),
-        verticalAlignment = Alignment.CenterVertically,
-
-
-        ) {
-        Icon(
-            painter = painterResource(id = R.drawable.ic_launcher_foreground),
-            contentDescription ="modderz logo",
-            modifier = Modifier.size(25.dp),
-            tint = Color.White
-        )
-
-        Text(
-            text = message,
-            color = Color.White,
-            fontSize = 15.sp,
-            textAlign = TextAlign.Center,
-        )
-
-    }
-}
